+++ flutterfast/lib/features/authentication/services/user_service/pocketbase_user_service.dart
@@ -1,0 +1,29 @@
+import 'package:flutterfast/features/authentication/models/fast_user.dart';
+import 'package:flutterfast/features/authentication/services/user_service/fast_user_service.dart';
+
+class PocketBaseUserService extends FastUserService {
+  @override
+  Future<void> createUser() {
+    // TODO: implement createUser
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> deleteUser(FastUser user) {
+    // TODO: implement deleteUser
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<FastUser?> getUser() {
+    // TODO: implement getUser
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> updateUser(FastUser user) {
+    // TODO: implement updateUser
+    throw UnimplementedError();
+  }
+
+}